fix: updating category

+++ src/main/java/com/hassan/store/dtos/CreateProductRequest.java
@@ -9,5 +9,5 @@ public class CreateProductRequest {
     private String name;
     private String description;
     private BigDecimal price;
-    private Long categoryId;
+    private Byte categoryId;
 }
